refactor(portal): parsing do acervo pela framework e DTOs em pasta própria
O Article carregava quatro parsers privados que reimplementavam o que o
fluent() já faz: text() com match sobre três tipos, number() com is_numeric,
parseDate() envolvendo Carbon::parse num try/catch e safeUrl() validando o
esquema da URL contra XSS. O objeto user aninhado ainda era desembrulhado à
mão antes de qualquer leitura.

Tudo isso vira acessor tipado do fluent(), inclusive a dot notation para
user.name e user.profile_image_90. A guarda de invariante — título e data,
sem os quais o artigo não tem lugar na ordenação nem na janela do destaque —
continua, agora em duas linhas. O DTO cai de 136 para 79 linhas.

Article, ArticleAuthor e ArticleTopic passam para Articles/DTOs/.

Comentários revisados. O docblock do ArticleFeed justificava ler da API com
um comando que não existe mais (devto:sync-articles); agora registra o estado
real: o contents já espelha o acervo em banco e migrar a leitura está
pendente. O log de descarte passa a dizer o motivo verdadeiro, falta de
título ou data, em vez de "payload inválido".
